Update code base
* [amebapro2][sdk] optimize the ai glass process

* [amebapro2][video] sync voe to 1.6.1.3

* [amebapro2][wlan] pno 2g use channel list

* [amebapro2][wlan] dynamic dtim

* [amebapro2][sdk] fix redefine compile warning

* [amebapro2][wlan] dynamic dtim

* [amebapro2][pmc] modify Sleep_PG and haL_Deep_Sleep

* [amebapro2][video] fix voe 1.6.1.3 fcs fail

* [amebapro2][sdk] add switch for version in uart protocal for ai glass

* [amebapro2][video] add timelapse example

* [amebapro2][mbedtls] add mbedtls-3.6.3

* [amebapro2][video] fix coverity issue

* [amebapro2][sdk] update multicast version ota in ai glass

* [amebapro2][sdk] split header and source files into dedicated folders in ai glass

* [amebapro2][wlan] add dhcps option to make cellphone switch to 4g/5g

+++ component/wifi/api/wifi_conf.h
@@ -1217,7 +1217,7 @@ int wifi_wowlan_set_wdt(u8  gpio,
  * @return  RTW_SUCCESS
  */
 int wifi_wowlan_set_bcn_track(u8  start_window,
-							  u8  max_window,
+							  u16  max_window,
 							  u8  increment_steps,
 							  u8  duration);
 #endif
